refactor(metrics): перенести расчётное ядро метрики брифа в reviewer/ (PRI-249)
