+++ eval/solve_task_metrics/__init__.py
@@ -4,7 +4,8 @@
 никогда не импортируется из reviewer/**. Расчётные модули (cost, classify,
 recall, history, forecast) — чистые функции без ввода-вывода.
 
-Когда за метрики возьмётся онлайн-версия (PRI-249), эти модули следует
-ПЕРЕНЕСТИ в продакшн-слой и импортировать в обе стороны, а не переписать
-второй раз: общий расчёт метрик не должен существовать в двух копиях.
+Расчётные модули (classify, recall, briefs) ПЕРЕНЕСЕНЫ в reviewer/metrics/
+brief_quality (PRI-249) и здесь только реэкспортируются: онлайн-съём метрики и
+этот харнесс обязаны мерить одной линейкой. Остальные модули (cost, ground_truth,
+history, snapshot, report, forecast, endtoend) офлайн-специфичны и живут здесь.
 """
