+++ reviewer/metrics/__init__.py
@@ -1,0 +1,1 @@
+"""Метрики качества самого reviewer (не метрики ревьюируемого кода)."""
