+++ reviewer/metrics/brief_quality/__init__.py
@@ -1,0 +1,7 @@
+"""Качество ретрива под бриф solve-task: core-recall, precision, таксономия промахов.
+
+Расчётное ядро общее для двух потребителей: онлайн-съёма на publish_review
+(PRI-249) и офлайн-харнесса eval/solve_task_metrics (PRI-250). Модули чистые —
+без ввода-вывода, без git, без БД: это условие того, что офлайн и онлайн меряют
+ОДНОЙ линейкой и числа «до/после» сравнимы.
+"""
